apps/partenaires/Serializers.py
from rest_framework import serializers
from .models import Partenaire

class PartenaireSerializer(serializers.ModelSerializer):
    class Meta:
        model = Partenaire
        fields = '__all__'

        